# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Listings Api",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/listings": {
            "get": {
                "tags": [
                    "Listings"
                ],
                "summary": "Получение всех объявлений",
                "operationId": "626511cff52fa1165beca4b7480a66c8",
                "parameters": [
                    {
                        "name": "paginate",
                        "in": "query",
                        "description": "Включить пагинацию",
                        "required": false,
                        "schema": {
                            "type": "bool"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список объявлений"
                    }
                }
            },
            "post": {
                "tags": [
                    "Listings"
                ],
                "summary": "Добавление объявления",
                "operationId": "f232854be0ba1c4e83eeda78909de489",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Название объявления",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Описание объявления",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "city_id",
                        "in": "query",
                        "description": "Id города",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "address",
                        "in": "query",
                        "description": "Адрес недвижимости",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "price_per_day",
                        "in": "query",
                        "description": "Цена за день аренды",
                        "required": true,
                        "schema": {
                            "type": "numeric"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Объявление создано"
                    },
                    "422": {
                        "description": "Ощибка валидации"
                    }
                }
            }
        },
        "/api/listings/{id}": {
            "get": {
                "tags": [
                    "Listings"
                ],
                "summary": "Получение объявления по id",
                "operationId": "3d6056b16aecf7f8e24443d7e69f9e1e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id объявления",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Объявление"
                    }
                }
            },
            "put": {
                "tags": [
                    "Listings"
                ],
                "summary": "Изменение названия и описания объявления",
                "operationId": "f0c3d97b00edab2226ce3a2e80b3c060",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id объявления",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Новое название",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Новое описание",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Объявление изменено"
                    },
                    "404": {
                        "description": "Объявление не найдено"
                    },
                    "403": {
                        "description": "Вы не владелец объявления"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Listings"
                ],
                "summary": "Изменение названия и описания объявления",
                "operationId": "9f2ef5353184b7ca889c8e4d150cc093",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id объявления",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Объявление изменено"
                    },
                    "404": {
                        "description": "Объявление не найдено"
                    }
                }
            }
        },
        "/api/user/messages": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Получение id пользователей, с которыми у вас есть переписка",
                "operationId": "ef914a0d0e80004d2a61bb6e5346b39d",
                "responses": {
                    "200": {
                        "description": "Список id пользователей"
                    }
                }
            },
            "put": {
                "tags": [
                    "Messages"
                ],
                "summary": "Изменение сообщения",
                "operationId": "509dac8f5c36cf23999cecb6c26d0b65",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id сообщения",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "text",
                        "in": "query",
                        "description": "Новый текст сообщения",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сообщение изменено"
                    },
                    "404": {
                        "description": "Сообщение не найдено"
                    },
                    "403": {
                        "description": "Вы не создатель сообщения"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Messages"
                ],
                "summary": "Удаление сообщения",
                "operationId": "b19161afa771d3337b26ac613f11eeeb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id сообщения",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сообщение удалено"
                    },
                    "404": {
                        "description": "Сообщение не найдено"
                    },
                    "403": {
                        "description": "Вы не создатель сообщения"
                    }
                }
            }
        },
        "/api/user/messages/{id}": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Получение переписки с пользователем",
                "operationId": "79b6efe1755ea718ded68cbc35355e24",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id получателя",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Все сообщения с сортировкой по времени"
                    }
                }
            },
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Создание сообщения пользователю с данным id",
                "operationId": "d54d6a23d19c8788988fe4dde6104079",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id получателя",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "text",
                        "in": "query",
                        "description": "Текст сообщения",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сообщение создано"
                    },
                    "404": {
                        "description": "Получатель не найден"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}